+++ Assets/Scriptables/Powers/OnDarkPowerSuffered.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ff5ea0183fa654648a45f7deaa2db378
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
