+++ Assets/Scripts/Player/PlayerObject.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: be2c2f73b47f4164d8de3d776e459830
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
