+++ Assets/Scripts/Power/Domain/IcePower.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 using Player;
 
-namespace Power.Domain
+namespace PowerModule.Domain
 {
     [CreateAssetMenu(fileName = "Ice", menuName = "Powers/Ice", order = 1)]
     public class IcePower : Power
