+++ Assets/Scripts/Power/Domain/ShieldPower.cs
@@ -3,17 +3,18 @@ using System.Collections;
 
 using Player;
 
-namespace Power.Domain
+namespace PowerModule.Domain
 {
     [CreateAssetMenu(fileName = "Shield", menuName = "Powers/Shield", order = 1)]
     public class ShieldPower : Power
     {
+
         [SerializeField]
         private float durationInSeconds;
 
         override public void onPowerReleased(PlayerObject player)
         {
-            // player.StartCoroutine(enableShieldCR(player));
+            player.StartCoroutine(enableShieldCR(player));
         }
         override public Power clone()
         {
@@ -25,11 +26,11 @@ namespace Power.Domain
             durationInSeconds = duration;
         }
 
-        // private IEnumerator enableShieldCR(PlayerObject player)
-        // {
-        //     player.turnOnShield(this, true);
-        //     yield return new WaitForSeconds(durationInSeconds);
-        //     player.turnOnShield(this, false);
-        // }
+        private IEnumerator enableShieldCR(PlayerObject player)
+        {
+            player.turnOnShield(true);
+            yield return new WaitForSeconds(durationInSeconds);
+            player.turnOnShield(false);
+        }
     }
 }
