+++ Assets/Scripts/Power/Events/OnDarkPowerSuffered.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 30ba2e5f09244f942accab84a3668c50
+guid: 45ea03dd4aea2be408750e968a0f2593
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
